用户信息接口 登出接口

+++ src/main/java/com/ligg/electronservice/pojo/Phone.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class Phone {
     private Integer phoneId;//id
-    private Integer phoneNumber;//手机号
+    private Long phoneNumber;//手机号
     private Integer lineStatus;//线路状态 1.在线、2.离线
     private String countryCode;//号码归属地
     private LocalDateTime registrationTime;//注册时间
